refactor: Move `is_writable()` into `.config`
- Change: Move `is_writable()` from `.utils` to `.config`.
  - Only used in `.config`.
